requirements.txt
openai-whisper
transformers
spacy
nltk
torch
streamlit
